Revert "Merge pull request #36 from Kylix-Finance/core-borrow-tests"
This reverts commit 6cb3706505e3314ed22395cd704dd91a9ec1b745, reversing
changes made to 913d1ca8da9714a82e527cb148643f19625afb74.
