+++ pallets/lending/src/lib.rs
@@ -79,7 +79,10 @@ mod borrow_repay;
 use borrow_repay::UserBorrow;
 
 #[cfg(test)]
-pub(crate) mod tests;
+mod mock;
+
+#[cfg(test)]
+mod tests;
 
 #[cfg(feature = "runtime-benchmarks")]
 mod benchmarking;
@@ -484,7 +487,7 @@ pub mod pallet {
 			Ok(())
 		}
 
-		fn update_borrow_index(&mut self) -> Result<(), Error<T>> {
+		fn udpate_borrow_index(&mut self) -> Result<(), Error<T>> {
 			let incr = self.calculate_compunded_interest()?;
 			let new_index =
 				self.borrow_index.checked_mul(&incr).ok_or(Error::<T>::OverflowError)?;
@@ -492,10 +495,10 @@ pub mod pallet {
 			Ok(())
 		}
 
-		pub fn update_indexes(&mut self) -> Result<(), Error<T>> {
+		fn update_indexes(&mut self) -> Result<(), Error<T>> {
 			if self.last_accrued_interest_at < Pallet::<T>::now_in_seconds() {
 				self.update_supply_index()?;
-				self.update_borrow_index()?;
+				self.udpate_borrow_index()?;
 			}
 			Ok(())
 		}
@@ -569,7 +572,8 @@ pub mod pallet {
 	/// StorageMap AssetPool { AssetId } => LendingPool { PoolId, Balance }
 	#[pallet::storage]
 	#[pallet::getter(fn reserve_pools)]
-	pub type LendingPoolStorage<T> = StorageMap<_, Blake2_128Concat, AssetPool<T>, LendingPool<T>>;
+	pub type LendingPoolStorage<T> =
+		StorageMap<_, Blake2_128Concat, AssetPool<T>, LendingPool<T>>;
 
 	// Now we need to define the properties of the underlying asset used in the lending pool
 	#[derive(
@@ -764,8 +768,8 @@ pub mod pallet {
 		/// * If the origin is not signed (i.e., the function was not called by a user).
 		/// * If the provided assets do not exist.
 		/// * If `amount` is 0 or less.
-		/// * If adding liquidity to the pool fails for any reason due to arithmetic overflow or
-		///   underflow
+		/// * If adding liquidity to the pool fails for any reason due to arithmetic overflows or
+		///   underflows
 		///
 		/// # Events
 		///
@@ -842,8 +846,8 @@ pub mod pallet {
 		/// * If the pool is not active.
 		/// * If the user has not enough liquidity to supply.
 		/// * If the balance amount to supply is not valid.
-		/// * If adding liquidity to the pool fails for any reason due to arithmetic overflow or
-		///  underflow
+		/// * If adding liquidity to the pool fails for any reason due to arithmetic overflows or
+		///  underflows
 		/// # Events
 		///
 		/// If the function succeeds, it triggers an event:
@@ -881,8 +885,8 @@ pub mod pallet {
 		/// * If the pool is not active.
 		/// * If the user has not enough liquidity to supply.
 		/// * If the balance amount to supply is not valid.
-		/// * If adding liquidity to the pool fails for any reason due to arithmetic overflow or
-		/// underflow
+		/// * If adding liquidity to the pool fails for any reason due to arithmetic overflows or
+		/// underflows
 		///
 		/// # Events
 		///
@@ -921,8 +925,8 @@ pub mod pallet {
 		/// * If the pool is not active.
 		/// * If the user has not enough liquidity to supply.
 		/// * If the balance amount to supply is not valid.
-		/// * If adding liquidity to the pool fails for any reason due to arithmetic overflow or
-		/// underflow
+		/// * If adding liquidity to the pool fails for any reason due to arithmetic overflows or
+		/// underflows
 		///
 		/// # Events
 		///
@@ -963,8 +967,8 @@ pub mod pallet {
 		/// * If the pool is not active.
 		/// * If the user has not enough liquidity to supply.
 		/// * If the balance amount to supply is not valid.
-		/// * If adding liquidity to the pool fails for any reason due to arithmetic overflow or
-		/// underflow
+		/// * If adding liquidity to the pool fails for any reason due to arithmetic overflows or
+		/// underflows
 		///
 		/// # Events
 		///
@@ -1501,7 +1505,7 @@ pub mod pallet {
 					&Self::account_id(),
 					who,
 					release_collateral_amount,
-					Preservation::Expendable,
+					Preservation::Preserve,
 				)?;
 			}
 
@@ -1573,7 +1577,7 @@ pub mod pallet {
 			Ok(())
 		}
 
-		/// Returns the block's timestamp in seconds as u64
+		/// Returns the the block's timestamp in seconds as u64
 		fn now_in_seconds() -> u64 {
 			core::time::Duration::from_millis(T::Time::now().saturated_into::<u64>())
 				.as_secs()
@@ -1621,28 +1625,26 @@ pub mod pallet {
 			Ok(amount)
 		}
 
-		/// Returns the amount of collateral asset to be released on partial repayment
+		/// Returns the amount of collateral asset to be released on partila repayement
 		/// Returns release_amount = pay / repayable_balance * collateral_balance
 		fn get_release_collateral_amount(
 			payment: AssetBalanceOf<T>,
 			total_due: AssetBalanceOf<T>,
 			collateral_balance: AssetBalanceOf<T>,
 		) -> Result<AssetBalanceOf<T>, Error<T>> {
-			let f_payment = FixedU128::from_inner(payment.saturated_into::<u128>());
-			let f_total_due = FixedU128::from_inner(total_due.saturated_into::<u128>());
-			let f_collateral_balance =
-				FixedU128::from_inner(collateral_balance.saturated_into::<u128>());
+			let f_payment = FixedU128::from(payment.saturated_into::<u128>());
+			let f_total_due = FixedU128::from(total_due.saturated_into::<u128>());
+			let f_collateral_balance = FixedU128::from(collateral_balance.saturated_into::<u128>());
 
-			let release_ratio =
-				f_payment.checked_div(&f_total_due).ok_or(Error::<T>::OverflowError)?;
-
-			let release_amount = release_ratio
+			let amount = f_payment
+				.checked_div(&f_total_due)
+				.ok_or(Error::<T>::OverflowError)?
 				.checked_mul(&f_collateral_balance)
 				.ok_or(Error::<T>::OverflowError)?
 				.into_inner()
 				.saturated_into();
 
-			Ok(release_amount)
+			Ok(amount)
 		}
 	}
 }
