--- pallets/lending/src/tests/mod.rs
@@ -1,3 +1,0 @@
-mod core_operations;
-pub(crate) mod mock;
-mod pool_management;
